requirements.txt
django==1.11
djangorestframework
djangorestframework-jwt
requests